- extract inner classes in their own files to ease up the integration process
