+++ src/main/java/com/thebuzzmedia/exiftool/ExifTool.java
@@ -16,14 +16,10 @@
 
 package com.thebuzzmedia.exiftool;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.List;
@@ -235,11 +231,11 @@ public class ExifTool {
 	 * Compiled {@link Pattern} of ": " used to split compact output from
 	 * ExifTool evenly into name/value pairs.
 	 */
-	private static final Pattern TAG_VALUE_PATTERN = Pattern
+	static final Pattern TAG_VALUE_PATTERN = Pattern
 			.compile("\\s*:\\s*");
-	private static final String STREAM_CLOSED_MESSAGE = "Stream closed";
+	static final String STREAM_CLOSED_MESSAGE = "Stream closed";
 
-	private static Logger log = LoggerFactory.getLogger(ExifTool.class);
+	static Logger log = LoggerFactory.getLogger(ExifTool.class);
 
 	/**
 	 * The absolute path to the ExifTool executable on the host system running
@@ -869,634 +865,5 @@ public class ExifTool {
 			}
 		}
 		return tagMap;
-	}
-
-	// ================================================================================
-	/**
-	 * Represents an Exif Process.
-	 */
-	public static class ExifProcess {
-		private final ReentrantLock closeLock = new ReentrantLock(false);
-		private final boolean keepAlive;
-		private final Process process;
-		private final BufferedReader reader;
-		private final OutputStreamWriter writer;
-		private volatile boolean closed = false;
-
-		public static VersionNumber readVersion(String exifCmd) {
-			ExifProcess process = _execute(false,
-					Arrays.asList(exifCmd, "-ver"));
-			try {
-				return new VersionNumber(process.readLine());
-			} catch (IOException ex) {
-				throw new RuntimeException(String.format(
-						"Unable to check version number of ExifTool: %s",
-						exifCmd));
-			} finally {
-				process.close();
-			}
-		}
-
-		public static Map<String, String> executeToResults(String exifCmd,
-				List<String> args) throws IOException {
-			List<String> newArgs = new ArrayList<String>(args.size() + 1);
-			newArgs.add(exifCmd);
-			newArgs.addAll(args);
-			ExifProcess process = _execute(false, newArgs);
-			try {
-				return process.readResponse();
-			} finally {
-				process.close();
-			}
-		}
-
-		public static String executeToString(String exifCmd, List<String> args)
-				throws IOException {
-			List<String> newArgs = new ArrayList<String>(args.size() + 1);
-			newArgs.add(exifCmd);
-			newArgs.addAll(args);
-			ExifProcess process = _execute(false, newArgs);
-			try {
-				return process.readResponseString();
-			} finally {
-				process.close();
-			}
-		}
-
-		public static ExifProcess startup(String exifCmd) {
-			List<String> args = Arrays.asList(exifCmd, "-stay_open", "True",
-					"-@", "-");
-			return _execute(true, args);
-		}
-
-		public static ExifProcess _execute(boolean keepAlive, List<String> args) {
-			log.debug(String
-					.format("Attempting to start external ExifTool process using args: %s",
-							args));
-			try {
-				Process process = new ProcessBuilder(args).start();
-				log.debug("\tSuccessful");
-				return new ExifProcess(keepAlive, process);
-			} catch (Exception e) {
-				String message = "Unable to start external ExifTool process using the execution arguments: "
-						+ args
-						+ ". Ensure ExifTool is installed correctly and runs using the command path '"
-						+ args.get(0)
-						+ "' as specified by the 'exiftool.path' system property.";
-
-				log.debug(message);
-				throw new RuntimeException(message, e);
-			}
-		}
-
-		public ExifProcess(boolean keepAlive, Process process) {
-			this.keepAlive = keepAlive;
-			this.process = process;
-			this.reader = new BufferedReader(new InputStreamReader(
-					process.getInputStream()));
-			this.writer = new OutputStreamWriter(process.getOutputStream());
-		}
-
-		public synchronized Map<String, String> sendArgs(List<String> args)
-				throws IOException {
-			StringBuilder builder = new StringBuilder();
-			for (String arg : args) {
-				builder.append(arg).append("\n");
-			}
-			builder.append("-execute\n");
-			writeFlush(builder.toString());
-			return readResponse();
-		}
-
-		public synchronized void writeFlush(String message) throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			writer.write(message);
-			writer.flush();
-		}
-
-		public synchronized String readLine() throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			return reader.readLine();
-		}
-
-		public synchronized Map<String, String> readResponse()
-				throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			log.debug("Reading response back from ExifTool...");
-			Map<String, String> resultMap = new HashMap<String, String>(500);
-			String line;
-
-			while ((line = reader.readLine()) != null) {
-				if (closed)
-					throw new IOException(STREAM_CLOSED_MESSAGE);
-				String[] pair = TAG_VALUE_PATTERN.split(line, 2);
-
-				if (pair.length == 2) {
-					resultMap.put(pair[0], pair[1]);
-					log.debug(String.format("\tRead Tag [name=%s, value=%s]",
-							pair[0], pair[1]));
-				}
-
-				/*
-				 * When using a persistent ExifTool process, it terminates its
-				 * output to us with a "{ready}" clause on a new line, we need
-				 * to look for it and break from this loop when we see it
-				 * otherwise this process will hang indefinitely blocking on the
-				 * input stream with no data to read.
-				 */
-				if (keepAlive && line.equals("{ready}")) {
-					break;
-				}
-			}
-			return resultMap;
-		}
-
-		public synchronized String readResponseString() throws IOException {
-			if (closed)
-				throw new IOException(STREAM_CLOSED_MESSAGE);
-			log.debug("Reading response back from ExifTool...");
-			String line;
-			StringBuilder result = new StringBuilder();
-			while ((line = reader.readLine()) != null) {
-				if (closed)
-					throw new IOException(STREAM_CLOSED_MESSAGE);
-
-				/*
-				 * When using a persistent ExifTool process, it terminates its
-				 * output to us with a "{ready}" clause on a new line, we need
-				 * to look for it and break from this loop when we see it
-				 * otherwise this process will hang indefinitely blocking on the
-				 * input stream with no data to read.
-				 */
-				if (keepAlive && line.equals("{ready}")) {
-					break;
-				} else
-					result.append(line);
-			}
-			return result.toString();
-		}
-
-		public boolean isClosed() {
-			return closed;
-		}
-
-		public void close() {
-			if (!closed) {
-				closeLock.lock();
-				try {
-					if (!closed) {
-						closed = true;
-						log.debug("Attempting to close ExifTool daemon process, issuing '-stay_open\\nFalse\\n' command...");
-						try {
-							writer.write("-stay_open\nFalse\n");
-							writer.flush();
-						} catch (IOException ex) {
-							// log.error(ex,ex);
-						}
-						try {
-							log.debug("Closing Read stream...");
-							reader.close();
-							log.debug("\tSuccessful");
-						} catch (Exception e) {
-							// no-op, just try to close it.
-						}
-
-						try {
-							log.debug("Closing Write stream...");
-							writer.close();
-							log.debug("\tSuccessful");
-						} catch (Exception e) {
-							// no-op, just try to close it.
-						}
-
-						log.debug("Read/Write streams successfully closed.");
-
-						try {
-							process.destroy();
-						} catch (Exception e) {
-							//
-						}
-						// process = null;
-
-					}
-				} finally {
-					closeLock.unlock();
-				}
-			}
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Enum used to define the different kinds of features in the native
-	 * ExifTool executable that this class can help you take advantage of.
-	 * <p/>
-	 * These flags are different from {@link Tag}s in that a "feature" is
-	 * determined to be a special functionality of the underlying ExifTool
-	 * executable that requires a different code-path in this class to take
-	 * advantage of; for example, <code>-stay_open True</code> support.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Feature {
-		/**
-		 * Enum used to specify that you wish to launch the underlying ExifTool
-		 * process with <code>-stay_open True</code> support turned on that this
-		 * class can then take advantage of.
-		 * <p/>
-		 * Required ExifTool version is <code>8.36</code> or higher.
-		 */
-		STAY_OPEN(8, 36);
-
-		private VersionNumber requireVersion;
-
-		private Feature(int... numbers) {
-			this.requireVersion = new VersionNumber(numbers);
-		}
-
-		/**
-		 * Used to get the version of ExifTool required by this feature in order
-		 * to work.
-		 * 
-		 * @return the version of ExifTool required by this feature in order to
-		 *         work.
-		 */
-		VersionNumber getVersion() {
-			return requireVersion;
-		}
-
-		boolean isSupported(VersionNumber exifVersionNumber) {
-			return requireVersion.isBeforeOrEqualTo(exifVersionNumber);
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Version Number used to determine if one version is after another.
-	 */
-	static class VersionNumber {
-		private final int[] numbers;
-
-		public VersionNumber(String str) {
-			String[] versionParts = str.trim().split("\\.");
-			this.numbers = new int[versionParts.length];
-			for (int i = 0; i < versionParts.length; i++) {
-				numbers[i] = Integer.parseInt(versionParts[i]);
-			}
-		}
-
-		public VersionNumber(int... numbers) {
-			this.numbers = numbers;
-		}
-
-		public boolean isBeforeOrEqualTo(VersionNumber other) {
-			int max = Math.min(this.numbers.length, other.numbers.length);
-			for (int i = 0; i < max; i++) {
-				if (this.numbers[i] > other.numbers[i]) {
-					return false;
-				} else if (this.numbers[i] < other.numbers[i]) {
-					return true;
-				}
-			}
-			// assume missing number as zero, so if the current process number
-			// is more digits it is a higher version
-			return this.numbers.length <= other.numbers.length;
-		}
-
-		@Override
-		public String toString() {
-			StringBuilder builder = new StringBuilder();
-			for (int i = 0; i < numbers.length; i++) {
-				if (builder.length() > 0) {
-					builder.append(".");
-				}
-				builder.append(numbers[i]);
-			}
-			return builder.toString();
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Enum used to define the 2 different output formats that {@link Tag}
-	 * values can be returned in: numeric or human-readable text.
-	 * <p/>
-	 * ExifTool, via the <code>-n</code> command line arg, is capable of
-	 * returning most values in their raw numeric form (e.g.
-	 * Aperture="2.8010323841") as well as a more human-readable/friendly format
-	 * (e.g. Aperture="2.8").
-	 * <p/>
-	 * While the {@link Tag}s defined on this class do provide a hint at the
-	 * type of the result (see {@link Tag#getType()}), that hint only applies
-	 * when the {@link Format#NUMERIC} form of the value is returned.
-	 * <p/>
-	 * If the caller finds the human-readable format easier to process,
-	 * {@link Format#HUMAN_READABLE} can be specified when calling
-	 * {@link ExifTool#getImageMeta(File, Format, Tag...)} and the returned
-	 * {@link String} values processed manually by the caller.
-	 * <p/>
-	 * In order to see the types of values that are returned when
-	 * {@link Format#HUMAN_READABLE} is used, you can check the comprehensive <a
-	 * href="http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html">
-	 * ExifTool Tag Guide</a>.
-	 * <p/>
-	 * This makes sense with some values like Aperture that in
-	 * {@link Format#NUMERIC} format end up returning as 14-decimal-place, high
-	 * precision values that are near the intended value (e.g.
-	 * "2.79999992203711" instead of just returning "2.8"). On the other hand,
-	 * other values (like Orientation) are easier to parse when their numeric
-	 * value (1-8) is returned instead of a much longer friendly name (e.g.
-	 * "Mirror horizontal and rotate 270 CW").
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Format {
-		NUMERIC,
-		HUMAN_READABLE
-	}
-
-	// ================================================================================
-	/**
-	 * Enum used to pre-define a convenient list of tags that can be easily
-	 * extracted from images using this class with an external install of
-	 * ExifTool.
-	 * <p/>
-	 * Each tag defined also includes a type hint for the parsed value
-	 * associated with it when the default {@link Format#NUMERIC} value format
-	 * is used.
-	 * <p/>
-	 * All replies from ExifTool are parsed as {@link String}s and using the
-	 * type hint from each {@link Tag} can easily be converted to the correct
-	 * data format by using the provided {@link Tag#parseValue(String)} method.
-	 * <p/>
-	 * This class does not make an attempt at converting the value automatically
-	 * in case the caller decides they would prefer tag values returned in
-	 * {@link Format#HUMAN_READABLE} format and to avoid any compatibility
-	 * issues with future versions of ExifTool if a tag's return value is
-	 * changed. This approach to leaving returned tag values as strings until
-	 * the caller decides they want to parse them is a safer and more robust
-	 * approach.
-	 * <p/>
-	 * The types provided by each tag are merely a hint based on the <a
-	 * href="http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html"
-	 * >ExifTool Tag Guide</a> by Phil Harvey; the caller is free to parse or
-	 * process the returned {@link String} values any way they wish.
-	 * <h3>Tag Support</h3>
-	 * ExifTool is capable of parsing almost every tag known to man (1000+), but
-	 * this class makes an attempt at pre-defining a convenient list of the most
-	 * common tags for use.
-	 * <p/>
-	 * This list was determined by looking at the common metadata tag values
-	 * written to images by popular mobile devices (iPhone, Android) as well as
-	 * cameras like simple point and shoots as well as DSLRs. As an additional
-	 * source of input the list of supported/common EXIF formats that Flickr
-	 * supports was also reviewed to ensure the most common/useful tags were
-	 * being covered here.
-	 * <p/>
-	 * Please email me or <a
-	 * href="https://github.com/thebuzzmedia/imgscalr/issues">file an issue</a>
-	 * if you think this list is missing a commonly used tag that should be
-	 * added to it.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Tag {
-		// single entry tags
-		APERTURE("ApertureValue", Double.class),
-		AUTHOR("XPAuthor", String.class),
-		COLOR_SPACE("ColorSpace", Integer.class),
-		COMMENT("XPComment", String.class),
-		CONTRAST("Contrast", Integer.class),
-		CREATE_DATE("CreateDate", String.class),
-		CREATION_DATE("CreationDate", String.class),
-		DATE_CREATED("DateCreated", String.class),
-		DATE_TIME_ORIGINAL("DateTimeOriginal", String.class),
-		DIGITAL_ZOOM_RATIO("DigitalZoomRatio", Double.class),
-		EXIF_VERSION("ExifVersion", String.class),
-		EXPOSURE_COMPENSATION("ExposureCompensation", Double.class),
-		EXPOSURE_PROGRAM("ExposureProgram", Integer.class),
-		EXPOSURE_TIME("ExposureTime", Double.class),
-		FLASH("Flash", Integer.class),
-		FOCAL_LENGTH("FocalLength", Double.class),
-		FOCAL_LENGTH_35MM("FocalLengthIn35mmFormat", Integer.class),
-		FNUMBER("FNumber", String.class),
-		GPS_ALTITUDE("GPSAltitude", Double.class),
-		GPS_ALTITUDE_REF("GPSAltitudeRef", Integer.class),
-		GPS_BEARING("GPSDestBearing", Double.class),
-		GPS_BEARING_REF("GPSDestBearingRef", String.class),
-		GPS_DATESTAMP("GPSDateStamp", String.class),
-		GPS_LATITUDE("GPSLatitude", Double.class),
-		GPS_LATITUDE_REF("GPSLatitudeRef", String.class),
-		GPS_LONGITUDE("GPSLongitude", Double.class),
-		GPS_LONGITUDE_REF("GPSLongitudeRef", String.class),
-		GPS_PROCESS_METHOD("GPSProcessingMethod", String.class),
-		GPS_SPEED("GPSSpeed", Double.class),
-		GPS_SPEED_REF("GPSSpeedRef", String.class),
-		GPS_TIMESTAMP("GPSTimeStamp", String.class),
-		IMAGE_HEIGHT("ImageHeight", Integer.class),
-		IMAGE_WIDTH("ImageWidth", Integer.class),
-		ISO("ISO", Integer.class),
-		KEYWORDS("XPKeywords", String.class),
-		LENS_MAKE("LensMake", String.class),
-		LENS_MODEL("LensModel", String.class),
-		MAKE("Make", String.class),
-		METERING_MODE("MeteringMode", Integer.class),
-		MODEL("Model", String.class),
-		ORIENTATION("Orientation", Integer.class),
-		OWNER_NAME("OwnerName", String.class),
-		RATING("Rating", Integer.class),
-		RATING_PERCENT("RatingPercent", Integer.class),
-		ROTATION("Rotation", Integer.class),
-		SATURATION("Saturation", Integer.class),
-		SENSING_METHOD("SensingMethod", Integer.class),
-		SHARPNESS("Sharpness", Integer.class),
-		SHUTTER_SPEED("ShutterSpeedValue", Double.class),
-		SOFTWARE("Software", String.class),
-		SUBJECT("XPSubject", String.class),
-		TITLE("XPTitle", String.class),
-		WHITE_BALANCE("WhiteBalance", Integer.class),
-		X_RESOLUTION("XResolution", Double.class),
-		Y_RESOLUTION("YResolution", Double.class),
-		// select ICC metadata
-		ICC_DESCRIPTION("ProfileDescription", String.class),
-		ICC_COLORSPACEDATA("ColorSpaceData", String.class),
-		// actually binary data, but what are we doing to do here??? Just use to
-		// save to file...
-		THUMBNAIL_IMAGE("ThumbnailImage", String.class),
-		THUMBNAIL_PHOTOSHOP("PhotoshopThumbnail", String.class), ;
-
-		/**
-		 * Used to get the {@link Tag} identified by the given, case-sensitive,
-		 * tag name.
-		 * 
-		 * @param name
-		 *            The case-sensitive name of the tag that will be searched
-		 *            for.
-		 * 
-		 * @return the {@link Tag} identified by the given, case-sensitive, tag
-		 *         name or <code>null</code> if one couldn't be found.
-		 */
-		public static Tag forName(String name) {
-			for (Tag tag : Tag.values()) {
-				if (tag.getName().equals(name)) {
-					return tag;
-				}
-			}
-			return null;
-		}
-
-		/**
-		 * Convenience method used to convert the given string Tag value
-		 * (returned from the external ExifTool process) into the type described
-		 * by the associated {@link Tag}.
-		 * 
-		 * @param <T>
-		 *            The type of the returned value.
-		 * @param value
-		 *            The {@link String} representation of the tag's value as
-		 *            parsed from the image.
-		 * 
-		 * @return the given string value converted to a native Java type (e.g.
-		 *         Integer, Double, etc.).
-		 * 
-		 * @throws IllegalArgumentException
-		 *             if <code>tag</code> is <code>null</code>.
-		 * @throws NumberFormatException
-		 *             if any exception occurs while trying to parse the given
-		 *             <code>value</code> to any of the supported numeric types
-		 *             in Java via calls to the respective <code>parseXXX</code>
-		 *             methods defined on all the numeric wrapper classes (e.g.
-		 *             {@link Integer#parseInt(String)} ,
-		 *             {@link Double#parseDouble(String)} and so on).
-		 * @throws ClassCastException
-		 *             if the type defined by <code>T</code> is incompatible
-		 *             with the type defined by {@link Tag#getType()} returned
-		 *             by the <code>tag</code> argument passed in. This class
-		 *             performs an implicit/unchecked cast to the type
-		 *             <code>T</code> before returning the parsed result of the
-		 *             type indicated by {@link Tag#getType()}. If the types do
-		 *             not match, a <code>ClassCastException</code> will be
-		 *             generated by the VM.
-		 */
-		@SuppressWarnings("unchecked")
-		public <T> T parseValue(String value) throws IllegalArgumentException {
-
-			if (value != null) {
-				if (Boolean.class.isAssignableFrom(this.getType())) {
-					return (T) Boolean.valueOf(value);
-				} else if (Byte.class.isAssignableFrom(this.getType())) {
-					return (T) Byte.valueOf(Byte.parseByte(value));
-				} else if (Integer.class.isAssignableFrom(this.getType())) {
-					return (T) Integer.valueOf(Integer.parseInt(value));
-				} else if (Short.class.isAssignableFrom(this.getType())) {
-					return (T) Short.valueOf(Short.parseShort(value));
-				} else if (Long.class.isAssignableFrom(this.getType())) {
-					return (T) Long.valueOf(Long.parseLong(value));
-				} else if (Float.class.isAssignableFrom(this.getType())) {
-					return (T) Float.valueOf(Float.parseFloat(value));
-				} else if (Double.class.isAssignableFrom(this.getType())) {
-					return (T) parseDouble(value);
-				} else if (Character.class.isAssignableFrom(this.getType())) {
-					return (T) Character.valueOf(value.charAt(0));
-				} else if (String.class.isAssignableFrom(this.getType())) {
-					return (T) value;
-				}
-			}
-			return null;
-		}
-
-		private Double parseDouble(String in) {
-			if (in.contains("/")) {
-				String[] enumeratorAndDivisor = in.split("/");
-				return Double.parseDouble(enumeratorAndDivisor[0])
-						/ Double.parseDouble(enumeratorAndDivisor[1]);
-			} else {
-				return Double.parseDouble(in);
-			}
-		}
-
-		/**
-		 * Used to get the name of the tag (e.g. "Orientation", "ISO", etc.).
-		 * 
-		 * @return the name of the tag (e.g. "Orientation", "ISO", etc.).
-		 */
-		public String getName() {
-			return name;
-		}
-
-		/**
-		 * Used to get a hint for the native type of this tag's value as
-		 * specified by Phil Harvey's <a href=
-		 * "http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html"
-		 * >ExifTool Tag Guide</a>.
-		 * 
-		 * @return a hint for the native type of this tag's value.
-		 */
-		public Class<?> getType() {
-			return type;
-		}
-
-		private String name;
-		private Class<?> type;
-
-		private Tag(String name, Class<?> type) {
-			this.name = name;
-			this.type = type;
-		}
-	}
-
-	// ================================================================================
-	public enum TagGroup {
-		EXIF("EXIF", "exif:all"),
-		IPTC("IPTC", "iptc:all"),
-		XMP("XMP", "xmp:all"),
-		ALL("ALL", "all"),
-		FILE("FILE", "file:all"),
-		ICC("ICC", "icc_profile:all");
-
-		private final String name;
-		private final String value;
-
-		private TagGroup(String name, String value) {
-			this.name = name;
-			this.value = value;
-		}
-
-		public String getName() {
-			return name;
-		}
-
-		public String getValue() {
-			return value;
-		}
-	}
-
-	// ================================================================================
-	/**
-	 * Class used to define an exception that occurs when the caller attempts to
-	 * use a {@link Feature} that the underlying native ExifTool install does
-	 * not support (i.e. the version isn't new enough).
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public class UnsupportedFeatureException extends RuntimeException {
-		private static final long serialVersionUID = -1332725983656030770L;
-
-		public UnsupportedFeatureException(Feature feature) {
-			super(
-					"Use of feature ["
-							+ feature.toString()
-							+ "] requires version "
-							+ feature.getVersion()
-							+ " or higher of the native ExifTool program. The version of ExifTool referenced by the system property 'exiftool.path' is not high enough. You can either upgrade the install of ExifTool or avoid using this feature to workaround this exception.");
-		}
-
 	}
 }
